first part format code

+++ src/main/java/currencyExchange/model/Statistic.java
@@ -4,11 +4,13 @@ import currencyExchange.enums.CurrencyType;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class Statistic {
-    private Double bought;
-    private Double sold;
-    private Double balance;
+    private BigDecimal bought;
+    private BigDecimal sold;
+    private BigDecimal balance;
     private CurrencyType currencyType;
 }
